Some logging fixes.
